discord_bot/requirements.txt
# PANDAUDIT Discord Bot Requirements

# Core Discord library
discord.py>=2.3.2

# For loading environment variables
python-dotenv>=1.0.0

# Optional: For enhanced logging
coloredlogs>=15.0.1

# Optional: For database support (if you want to persist warnings/mutes)
# Uncomment if needed:
# sqlalchemy>=2.0.0
# aiosqlite>=0.19.0

# Optional: For web scraping (to fetch latest blog posts)
# Uncomment if needed:
# aiohttp>=3.9.0
# beautifulsoup4>=4.12.0
